Refactor IoUringManager into .h/.cpp, add SyncIoManager fallback class
- Split IoUringManager from header-only into .h/.cpp
- Extract synchronous pread fallback into a separate SyncIoManager class
  with the same BatchHandle/addBatch/wait interface
- IoUringManager is now a pure io_uring class with a single #ifdef wrapping
  it from outside (no scattered #ifdef inside the class body)
- Add BatchIoManager alias selecting IoUringManager or SyncIoManager at
  compile time depending on QLEVER_HAS_IO_URING
- Add IoUringManager.cpp to the util library in src/util/CMakeLists.txt
- Convert tests to TYPED_TEST, exercising both classes when io_uring is
  present (14 tests total: 7 per type)
